JET-323: allow free movement for deposit notes
Allow users of the program to have their deposit notes transferred to
token accounts outside the program control.

Introduces two new instructions to facilitate this
  * `deposit_tokens` - like `deposit`, but destination account for notes
    can be any token account.
  * `withdraw_tokens` - like `withdraw`, but source account for notes
    can be any token account.
